dynamically generate list of messages for testing (#117)
 🐿 v2.10.3

Whilst yak-shaving on another issue on this repo yesterday i noticed the list of features for pa11y testing was hardcoded. As this list changes pretty frequently and the tests are somewhat tucked away (i.e. easy to forget to update) I thought it would be better to generate the list dynamically from the manifest which _should_ be representative of reality