Update docs to include info about linting
This is to avoid the issues that I had along with other devs from my team when trying to commit to the repo. While consumer products devs may all have ES Lint and Editor Config, not every team does, so clarifying that these are needed to be able to see errors when contributing is very helpful.

Additional changes have been made to the n-gage repo to make seeing and fixing linting errors easier.