Remove alt text for banner images
Currently if a banner title exists, it is used as the alt text for the
banner image. This may cause issues for screen reader users because the
image is placed next to the text so the message is announced twice.

Also, the text itself doesn't convey meaningful information about the
image.

Trello ticket: https://trello.com/c/i2TjYybE/154-decorative-image-with-alt-issue-id-dacdecorativeimage01 🐿 v2.12.5