Remove unnecessary getters from slot presenter.
For some reason, handlebars was calling the presenter (I was able to get
a breakpoint to catch inside the getters) but wasn't receiving the value
being returned. Eventually came to the realisation that these don't need
to be getters at all as far as we could tell so just setting them as
static properties. 🐿 v2.12.5